+++ src/ResourceManager/Websites/Commands.Websites/Cmdlets/WebApps/SetAzureWebApp.cs
@@ -125,10 +125,8 @@ namespace Microsoft.Azure.Commands.WebApps.Cmdlets.WebApps
         [Parameter(ParameterSetName = ParameterSet1Name, Mandatory = false, HelpMessage = "Enable/disable redirecting all traffic to HTTPS on an existing azure webapp")]
         public bool HttpsOnly { get; set; }
 
-        [Parameter(ParameterSetName = ParameterSet1Name, Mandatory = false, HelpMessage = "Web app Azure Storage account path. Only supported in Windows Containers and Linux Web apps. Example: -AzureStorageAccounts @{ AzureStorageAccount1 = @{ Type = \"AzureFiles\"; AccountName = \"My Azure Storage account name\"; ShareName = \"My Azure Storage account share name\"; AccessKey = \"My Azure Storage account access key\"; MountPath = \"The path to mount inside the container. For example: C:\\myfolder for Windows Containers or /myfolder for Linux Web apps\"}; AzureStorageAccount2 = @{ Type = \"AzureFiles\"; AccountName = \"My Azure Storage account name\"; ShareName = \"My Azure Storage account share name\"; AccessKey = \"My Azure Storage account access key\"; MountPath = \"The path to mount inside the container. For example: C:\\myfolder for Windows Containers or /myfolder for Linux Web apps\"} }")]
-        [ValidateNotNullOrEmpty]
-        [ValidateAzureStorageAccounts]
-        public Hashtable AzureStorageAccountPath { get; set; }
+        [Parameter(ParameterSetName = ParameterSet1Name, Mandatory = false, HelpMessage = "Azure Storage to mount inside a Web App for Container. Use New-AzureRmWebAppAzureStoragePath to create it")]
+        public WebAppAzureStoragePath[] AzureStoragePath { get; set; }
 
         public override void ExecuteCmdlet()
         {
@@ -223,7 +221,8 @@ namespace Microsoft.Azure.Commands.WebApps.Cmdlets.WebApps
                     }
 
                     // Update web app configuration
-                    WebsitesClient.UpdateWebAppConfiguration(ResourceGroupName, location, Name, null, siteConfig, appSettings.ConvertToStringDictionary(), ConnectionStrings.ConvertToConnectionStringDictionary(), AzureStorageAccountPath.ConvertToAzureStorageAccountPathPropertyDictionary());
+
+                    WebsitesClient.UpdateWebAppConfiguration(ResourceGroupName, location, Name, null, siteConfig, appSettings.ConvertToStringDictionary(), ConnectionStrings.ConvertToConnectionStringDictionary(), AzureStoragePath.ConvertToAzureStorageAccountPathPropertyDictionary());
 
                     //Update WebApp object after configuration update
                     WebApp = new PSSite(WebsitesClient.GetWebApp(ResourceGroupName, Name, null));