+++ src/ResourceManager/Websites/Commands.Websites/ChangeLog.md
@@ -20,7 +20,7 @@
 ## Current Release
 * New Cmdlet Get-AzureRMWebAppContainerContinuousDeploymentUrl - Gets the Container Continuous Deployment Webhook URL
 * New Cmdlets New-AzureRMWebAppContainerPSSession and Enter-WebAppContainerPSSession  - Initiates a PowerShell remote session into a windows container app
-* Set-AzureRmWebApp and Set-AzureRmWebAppSlot - New parameter (-AzureStorageAccounts) added to specify Azure Storage accounts to be mounted in Windows and Linux container apps
+* Set-AzureRmWebApp and Set-AzureRmWebAppSlot - New parameter (-AzureStoragePath) added to specify Azure Storage paths to be mounted in Windows and Linux container apps. Use the output of the new cmdlet New-AzureRmWebAppAzureStoragePath as a parameter to set the Azure Storage paths.
 
 ## Version 5.1.0
 * Updating to use the latest .NET SDK version (2.0.0)